requirements.txt
pydantic-ai==0.0.42
numpy==2.2.4
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
tavily-python==0.5.1
langchain-google-genai==2.1.4
langchain==0.3.25
langgraph==0.4.3
langchain-community==0.3.24
langchain-openai==0.3.16